+++ tests/unit/models/v0.3/modelCard.test.js
@@ -28,6 +28,7 @@ const basicCard = {
   username: "username",
   pin: "1234",
   // required for web applicants
+  email: "test@test.com",
   birthdate: "01/01/1988",
   acceptTerms: true,
   ageGate: true,
@@ -847,14 +848,22 @@ describe("Card", () => {
         "email cannot be empty for this policy type.",
       );
     });
-    it("should fail for webApplicant policies without a birthdate", async () => {
+    it("should fail for webApplicant policies without a birthdate or email", async () => {
       const cardNoEmail = new Card({
+        ...basicCard,
+        email: undefined,
+        policy: Policy({ policyType: "webApplicant" }),
+      });
+      const cardNoBirthdate = new Card({
         ...basicCard,
         birthdate: undefined,
         policy: Policy({ policyType: "webApplicant" }),
       });
 
       await expect(cardNoEmail.validate()).rejects.toThrow(
+        "email cannot be empty for this policy type.",
+      );
+      await expect(cardNoBirthdate.validate()).rejects.toThrow(
         "birthdate cannot be empty for this policy type.",
       );
     });
@@ -1096,8 +1105,8 @@ describe("Card", () => {
         policy: webApplicant,
       });
 
-      expect(card.requiredByPolicy("email")).toEqual(false);
-      expect(card.requiredByPolicy("barcode")).toEqual(false);
+      expect(card.requiredByPolicy("barcode")).toEqual(true);
+      expect(card.requiredByPolicy("email")).toEqual(true);
       expect(card.requiredByPolicy("birthdate")).toEqual(true);
       expect(card.requiredByPolicy("ageGate")).toEqual(false);
     });
@@ -1108,7 +1117,7 @@ describe("Card", () => {
         policy: simplyeJuvenile,
       });
 
-      expect(card.requiredByPolicy("email")).toEqual(false);
+      expect(card.requiredByPolicy("email")).toEqual(true);
       expect(card.requiredByPolicy("barcode")).toEqual(true);
       expect(card.requiredByPolicy("birthdate")).toEqual(false);
     });
@@ -1136,7 +1145,7 @@ describe("Card", () => {
       "soLicenseKey",
     );
 
-    it("always returns false for web applications with or without a work address", () => {
+    it("returns false for web applications without a work address", () => {
       let card = new Card({
         ...basicCard,
         policy: webApplicant,
@@ -1155,7 +1164,7 @@ describe("Card", () => {
         workAddress: workAddressInCity,
         policy: webApplicant,
       });
-      expect(card.worksInCity()).toEqual(false);
+      expect(card.worksInCity()).toEqual(true);
     });
 
     it("returns false because there is no work address", () => {
@@ -1190,14 +1199,6 @@ describe("Card", () => {
   describe("livesOrWorksInCity", () => {
     const simplyePolicy = Policy();
     const webApplicant = Policy({ policyType: "webApplicant" });
-    const addressNotInCity = new Address(
-      {
-        line1: "street address",
-        city: "Albany",
-        state: "New York",
-      },
-      "soLicenseKey",
-    );
 
     const addressInCity = new Address(
       {
@@ -1208,37 +1209,19 @@ describe("Card", () => {
       "soLicenseKey",
     );
 
-    it("always returns false for web applications", () => {
-      let card = new Card({
-        ...basicCard,
-        policy: webApplicant,
-      });
-      expect(card.livesOrWorksInCity()).toEqual(false);
-
-      // Doesn't matter if they have a work address...
-      card = new Card({
-        ...basicCard,
-        workAddress: addressNotInCity,
-        policy: webApplicant,
-      });
-      expect(card.livesOrWorksInCity()).toEqual(false);
-
-      // even if that work address is in the city.
-      card = new Card({
-        ...basicCard,
-        workAddress: addressInCity,
-        policy: webApplicant,
-      });
-      expect(card.livesOrWorksInCity()).toEqual(false);
-    });
-
     it("returns false because they do not live in NYC", () => {
       const card = new Card({
         ...basicCard,
         address: new Address({ city: "Albany" }, "soLicenseKey"),
         policy: simplyePolicy,
       });
+      const cardWeb = new Card({
+        ...basicCard,
+        address: new Address({ city: "Albany" }, "soLicenseKey"),
+        policy: webApplicant,
+      });
       expect(card.livesOrWorksInCity()).toEqual(false);
+      expect(cardWeb.livesOrWorksInCity()).toEqual(false);
     });
 
     it("returns true because they do not live in NYC but work there", () => {
@@ -1248,7 +1231,14 @@ describe("Card", () => {
         workAddress: addressInCity,
         policy: simplyePolicy,
       });
+      const cardWeb = new Card({
+        ...basicCard,
+        address: new Address({ city: "Albany" }, "soLicenseKey"),
+        workAddress: addressInCity,
+        policy: webApplicant,
+      });
       expect(card.livesOrWorksInCity()).toEqual(true);
+      expect(cardWeb.livesOrWorksInCity()).toEqual(true);
     });
 
     it("returns true if they live in NYC regardless of work address", () => {
@@ -1264,6 +1254,19 @@ describe("Card", () => {
         policy: simplyePolicy,
       });
       expect(card.livesOrWorksInCity()).toEqual(true);
+
+      let cardWeb = new Card({
+        ...basicCard,
+        policy: webApplicant,
+      });
+      expect(cardWeb.livesOrWorksInCity()).toEqual(true);
+
+      cardWeb = new Card({
+        ...basicCard,
+        workAddress: addressInCity,
+        policy: webApplicant,
+      });
+      expect(cardWeb.livesOrWorksInCity()).toEqual(true);
     });
   });
 
@@ -1274,7 +1277,7 @@ describe("Card", () => {
     const addressNotNY = new Address({ state: "New Jersey" }, "soLicenseKey");
     const addressNY = new Address({ state: "New York" }, "soLicenseKey");
 
-    it("returns false for all web applicants if they are in NY state or not", () => {
+    it("returns false for web applicants if they are not in NY state", () => {
       const cardNotNY = new Card({
         ...basicCard,
         address: addressNotNY,
@@ -1287,10 +1290,10 @@ describe("Card", () => {
       });
 
       expect(cardNotNY.livesInState()).toEqual(false);
-      expect(cardNY.livesInState()).toEqual(false);
+      expect(cardNY.livesInState()).toEqual(true);
     });
 
-    it("returns false if they are not in NY state", () => {
+    it("returns false for simplye applicants if they are not in NY state", () => {
       const cardNotNY = new Card({
         ...basicCard,
         address: addressNotNY,
@@ -1549,10 +1552,11 @@ describe("Card", () => {
         policy: Policy({ policyType: "webApplicant" }),
       });
 
-      // A temporary or standard card has an expiration of 90 days.
+      // A standard card has an expiration of 1095 days (3 years).
       expect(card.isTemporary).toEqual(false);
-      expect(card.getExpirationDays()).toEqual(90);
+      expect(card.getExpirationDays()).toEqual(1095);
 
+      // A temporary card has an expiration of 90 days.
       card.isTemporary = true;
       expect(card.getExpirationDays()).toEqual(90);
     });
@@ -1682,7 +1686,7 @@ describe("Card", () => {
     const simplyePolicy = Policy({ policyType: "simplye" });
     const addressNotNY = new Address({ city: "Hoboken", state: "New Jersey" });
 
-    it("returns a temporary card for web applicants", () => {
+    it("returns a temporary card for web applicants out of NYS or not residential but in NYS", () => {
       const card = new Card({
         ...basicCard,
         address: new Address({
@@ -1700,11 +1704,27 @@ describe("Card", () => {
 
       expect(card.getCardType()).toEqual({
         ...Card.RESPONSES.temporaryCard,
-        reason: "The policy for this card is web applicant.",
+        reason: "The home address is in NYC but is not residential.",
       });
       expect(cardNotNY.getCardType()).toEqual({
         ...Card.RESPONSES.temporaryCard,
-        reason: "The policy for this card is web applicant.",
+        reason:
+          "The home address is not in New York State but the work address is in New York City.",
+      });
+    });
+
+    it("returns a standard card for web applicants in NYS and residential", () => {
+      const cardResidential = new Card({
+        ...basicCard,
+        address: new Address({
+          city: "New York",
+          state: "New York",
+          isResidential: "true",
+        }),
+        policy: Policy({ policyType: "webApplicant" }),
+      });
+      expect(cardResidential.getCardType()).toEqual({
+        ...Card.RESPONSES.standardCard,
       });
     });
 
@@ -1849,7 +1869,7 @@ describe("Card", () => {
       );
     });
 
-    it("does not attempt to create a barcode for web applicants", async () => {
+    it("attempts to create a barcode for web applicants", async () => {
       IlsClient.mockImplementation(() => ({
         createPatron: () => Promise.resolve({ data: { link: "some patron" } }),
       }));
@@ -1865,10 +1885,10 @@ describe("Card", () => {
       const spy = jest.spyOn(card, "setBarcode");
 
       await card.createIlsPatron();
-      expect(spy).not.toHaveBeenCalled();
+      expect(spy).toHaveBeenCalled();
     });
 
-    it("does attempt to create a barcode for simplye applicants", async () => {
+    it("attempts to create a barcode for simplye applicants", async () => {
       IlsClient.mockImplementation(() => ({
         createPatron: () => Promise.resolve({ data: { link: "some patron" } }),
       }));
@@ -1892,7 +1912,7 @@ describe("Card", () => {
       expect(spy).toHaveBeenCalled();
     });
 
-    it("does attempt to create a barcode but fails!", async () => {
+    it("attempts to create a barcode but fails!", async () => {
       IlsClient.mockImplementation(() => ({
         createPatron: () => Promise.resolve({ data: { link: "some patron" } }),
       }));
@@ -1945,7 +1965,6 @@ describe("Card", () => {
       expect(spy).toHaveBeenCalled();
     });
 
-    // TODO:
     it("creates a patron", async () => {
       // Mock that the ILS fails
       IlsClient.mockImplementation(() => ({
@@ -2061,7 +2080,7 @@ describe("Card", () => {
       );
     });
 
-    it("returns a temporary card for web applicants", () => {
+    it("returns a temporary card for web applicants not in NYS", () => {
       const card = new Card({
         ...basicCard,
         address: new Address({
@@ -2074,7 +2093,7 @@ describe("Card", () => {
       card.isTemporary = true;
       card.cardType = card.getCardType();
       expect(card.selectMessage()).toEqual(
-        "The library card will be a temporary library card. The policy for this card is web applicant. Visit your local NYPL branch within 90 days to upgrade to a standard card.",
+        "The library card will be a temporary library card. The home address is not in New York State but the work address is in New York City. Visit your local NYPL branch within 90 days to upgrade to a standard card.",
       );
     });